Dataflow analysis for shape inference
Original review: https://github.com/ftynse/water/pull/20

Introduce a dataflow-based pass for inferring shapes of tensors in the
wave dialect. The pass simultaneously performs sparse forward and
backward dataflow analysis propagating known shapes of wave tensors
across operations. Details of propagation for wave dialect operations
are deferred to each operation via an interface letting it define
forward and backward transfer functions. A couple of traits are
predefined with default implementation of the interface methods for most
common cases.

Wrapping the process in the dataflow framework provides convergence and
soundness guarantees assuming transfer functions are monotonous and
sound. This in turn allows one to easily detect and report shape
conflicts in the dialect.

Signed-off-by: Alex Zinenko <git@ozinenko.com>