Verify usage of hyperparameters
Original review: https://github.com/ftynse/water/pull/46

When hyperparams are present, verify that any symbol used inside result
or block argument types is defined in the list of hyperparameters. This
ensures the IR can be lowered and correctly processed by passes.

We currently don't have a mechanism to find all other uses of symbols,
for example those inside attributes attached to ops. We may need an
interface to support that.

---------

Signed-off-by: Alex Zinenko <git@ozinenko.com>
Signed-off-by: tyb0807 <sontuan.vu@amd.com>
Co-authored-by: tyb0807 <sontuan.vu@amd.com>